% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{coastlines}
\alias{coastlines}
\alias{coastlines110}
\alias{coastlines50}
\alias{coastline110}
\alias{coastline50}
\title{world coastlines from Natural Earth}
\format{
An object of class `sf`

An object of class \code{sf} (inherits from \code{data.frame}) with 134 rows and 4 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1428 rows and 4 columns.
}
\source{
\url{https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/110m/physical/ne_110m_coastline.zip}

\url{https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/50m/physical/ne_50m_coastline.zip}
}
\usage{
coastline110

coastline50
}
\description{
coastline lines at scales 1:110m (small), 1:50m (medium). 1:10m (large) are in the package rnaturalearthhires.
}
\keyword{datasets}
