% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{map_units}
\alias{map_units}
\alias{map_units110}
\alias{map_units50}
\title{world map_unit polygons from Natural Earth}
\format{
A \code{SpatialPolygonsDataFrame}

An object of class \code{sf} (inherits from \code{data.frame}) with 183 rows and 169 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 265 rows and 169 columns.
}
\source{
\url{http//www.naturalearthdata.com/download/110m/cultural/ne_110m_admin_0_map_units.zip}

\url{http//www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_0_map_units.zip}
}
\usage{
map_units110

map_units50
}
\description{
map_unit polygons at scales 1:110m (small), 1:50m (medium). 1:10m (large) are in the package rnaturalearthhires.
There are more map_units than countries e.g. United Kingdom is split into England, Scotland, Wales and Northern Ireland.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with attributes}
}}

\keyword{datasets}
