% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaturalearthdata-package.R
\docType{package}
\name{rnaturalearthdata-package}
\alias{rnaturalearthdata}
\alias{rnaturalearthdata-package}
\title{rnaturalearthdata: World Vector Map Data from Natural Earth Used in 'rnaturalearth'}
\description{
Vector map data from \url{https://www.naturalearthdata.com/}. Access functions are provided in the accompanying package 'rnaturalearth'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/rnaturalearthdata/}
  \item \url{https://github.com/ropensci/rnaturalearthdata}
  \item Report bugs at \url{https://github.com/ropensci/rnaturalearthdata/issues}
}

}
\author{
\strong{Maintainer}: Philippe Massicotte \email{pmassicotte@hotmail.com} (\href{https://orcid.org/0000-0002-5919-4116}{ORCID})

Authors:
\itemize{
  \item Andy South \email{southandy@gmail.com} (\href{https://orcid.org/0000-0003-4051-6135}{ORCID})
  \item Schramm Michael \email{mpschramm@gmail.com}
}

}
\keyword{internal}
