% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaturalearthdata-package.r
\docType{package}
\name{rnaturalearthdata}
\alias{rnaturalearthdata}
\alias{rnaturalearthdata-package}
\title{rnaturalearthdata : world vector maps from Natural Earth}
\description{
Facilitates world mapping by making \href{http://www.naturalearthdata.com/}{Natural Earth} map data more easily available to R users.
}
\seealso{
\code{\link[rnaturalearth]{ne_countries}} \code{\link[rnaturalearth]{ne_states}} \code{\link[rnaturalearth]{ne_download}}
}

