% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sovereignty}
\alias{sovereignty}
\alias{sovereignty110}
\alias{sovereignty50}
\title{world sovereignty polygons from Natural Earth}
\format{
An object of class `sf`

An object of class \code{sf} (inherits from \code{data.frame}) with 171 rows and 169 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 201 rows and 169 columns.
}
\source{
\url{https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/110m/cultural/ne_110m_admin_0_sovereignty.zip}

\url{https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_0_sovereignty.zip}
}
\usage{
sovereignty110

sovereignty50
}
\description{
sovereignty polygons at scales 1:110m (small), 1:50m (medium). 1:10m (large) are in the package rnaturalearthhires.
There are fewer sovereign states than countries e.g. the Falkland Islands are included with the United Kingdom.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with attributes}
}}

\keyword{datasets}
