% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{states}
\alias{states}
\alias{states50}
\alias{states50}
\title{state (admin level 1) polygons from Natural Earth}
\format{A \code{SpatialPolygonsDataFrame}}
\source{
\url{http//www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_1_states.zip}
}
\usage{
states50
}
\description{
For Australia, Brazil, Canada and USA, at 1:50m (medium) resolution. The package rnaturalearthhires has all world countries at 1:10m (high) resolution.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with attributes}
}}

\keyword{datasets}
