% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaturalearthdata-package.r
\docType{package}
\name{rnaturalearthdata}
\alias{rnaturalearthdata}
\alias{rnaturalearthdata-package}
\title{rnaturalearthdata : World Vector Map Data from Natural Earth Used in 'rnaturalearth'}
\description{
holds data used by rnaturalearth package from \href{http://www.naturalearthdata.com/}{Natural Earth}.
}
