% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tiny_countries}
\alias{tiny_countries}
\alias{tiny_countries110}
\alias{tiny_countries50}
\title{tiny countries points from Natural Earth}
\format{
An object of class `sf`

An object of class \code{sf} (inherits from \code{data.frame}) with 37 rows and 171 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 76 rows and 171 columns.
}
\source{
\url{http//www.naturalearthdata.com/download/110m/physical/ne_110m_tiny_countries.zip}

\url{http//www.naturalearthdata.com/download/50m/physical/ne_50m_tiny_countries.zip}
}
\usage{
tiny_countries110

tiny_countries50
}
\description{
small country points at scales 1:110m (small), 1:50m (medium).
}
\keyword{datasets}
