% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{tiny_countries}
\alias{tiny_countries}
\alias{tiny_countries110}
\alias{tiny_countries50}
\title{tiny countries points from Natural Earth}
\format{A \code{SpatialPointsDataFrame}}
\source{
\url{http//www.naturalearthdata.com/download/110m/physical/ne_110m_tiny_countries.zip}

\url{http//www.naturalearthdata.com/download/50m/physical/ne_50m_tiny_countries.zip}
}
\usage{
tiny_countries110

tiny_countries50
}
\description{
small country points at scales 1:110m (small), 1:50m (medium).
}
\keyword{datasets}

